Giving up on DerivingVia. Got default signatures working instead!
The reason I couldn't get DerivingVia to work is that I couldn't get
the type parameter to AnyK to have type role 'representational'. I
tried rewriting it, but couldn't come up with any way to achieve it.

Also usage of deriveGenericK in combination with GenericKWrapper2
causes compiler hang, but this more direct approach bypasses whatever
is going on there.

So, for now, this is a clearly superior approach for automatically
deriving DepKDeserialize.